split up some value_counts groupby tests a bit
